// generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  CUSTOMER: 'CUSTOMER',
  PROVIDER: 'PROVIDER',
  ADMIN: 'ADMIN'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const UserStatus = {
  ACTIVE: 'ACTIVE',
  SUSPENDED: 'SUSPENDED'
} as const

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus]


export const RentalStatus = {
  PLACED: 'PLACED',
  CONFIRMED: 'CONFIRMED',
  CANCELLED: 'CANCELLED',
  PAID: 'PAID',
  PICKED_UP: 'PICKED_UP',
  RETURNED: 'RETURNED'
} as const

export type RentalStatus = (typeof RentalStatus)[keyof typeof RentalStatus]


export const PaymentProvider = {
  STRIPE: 'STRIPE',
  SSLCOMMERZ: 'SSLCOMMERZ'
} as const

export type PaymentProvider = (typeof PaymentProvider)[keyof typeof PaymentProvider]


export const PaymentStatus = {
  PENDING: 'PENDING',
  COMPLETED: 'COMPLETED',
  FAILED: 'FAILED'
} as const

export type PaymentStatus = (typeof PaymentStatus)[keyof typeof PaymentStatus]